updated mfe.py to handle entropy correction, and scale charge. closes #93
