+++ testcase_1A2P/run.prm
@@ -1,8 +1,24 @@
 /home/jmao/projects/Stable-MCCE    (MCCE_HOME)
-t                       (DO_ANALYSIS)
-t                       (GET_HBOND_MATRIX)
+t                       (DO_MONTE)
 /home/jmao/projects/Stable-MCCE/extra.tpl    (EXTRA)
-1.2                     (HBOND_LOWER_LIMIT)
-3.2                     (HBOND_UPPER_LIMIT)
-90.0                    (HBOND_ANG_CUTOFF)
-t                       (GET_HBOND_NETWORK)
+ph                      (TITR_TYPE)
+0.0                     (TITR_PH0)
+0.0                     (TITR_EH0)
+1.0                     (TITR_PHD)
+1.0                     (TITR.EHD)
+15                      (TITR_STEPS)
+5.0                     (BIG_PAIRWISE)
+-1                      (MONTE_SEED)
+298.15                  (MONTE_T)
+3                       (MONTE_FLIPS)
+100                     (MONTE_NSTART)
+300                     (MONTE_NEQ)
+0.001                   (MONTE_REDUCE)
+6                       (MONTE_RUNS)
+2000                    (MONTE_NITER)
+50000                   (MONTE_TRACE)
+1000000                 (NSTATE_MAX)
+f                       (MONTE_TSX)
+f                       (MFE_POINT)
+-1.0                    (MFE_CUTOFF)
+f                       (MS_OUT)
